requirements.txt
# Legal Contract Assistant - Dependencies

# Web Framework
flask==3.0.0
flask-cors==4.0.0

# Environment Variables
python-dotenv==1.0.0

# NLP & Text Processing
spacy==3.7.2
nltk==3.8.1

# Document Parsing
PyPDF2==3.0.1
python-docx==1.1.0

# LLM Integration
groq==0.31.0

# PDF Generation
reportlab==4.0.7

# Language Detection
langdetect==1.0.9

# Utilities
werkzeug==3.0.1

# Note: After installing, download spaCy models:
# python -m spacy download en_core_web_lg
# python -m spacy download xx_ent_wiki_sm
